test(angular): add new pages to navigate between lazy and standalone tests
