+++ packages/angular/test/base/src/app/app-landing/app-landing.component.html
@@ -1,0 +1,16 @@
+<ion-header>
+  <ion-toolbar>
+    <ion-title>Test App - Angular {{ angularVersion.major }} </ion-title>
+  </ion-toolbar>
+</ion-header>
+
+<ion-content>
+  <ion-list>
+    <ion-item routerLink="/lazy">
+      <ion-label>Go to Lazy App</ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone">
+      <ion-label>Go to Standalone App</ion-label>
+    </ion-item>
+  </ion-list>
+</ion-content>
