+++ packages/angular/test/base/src/app/app-landing/app-landing.component.ts
@@ -1,0 +1,13 @@
+import { Component, VERSION } from '@angular/core';
+
+@Component({
+  selector: 'app-landing',
+  templateUrl: './app-landing.component.html',
+  standalone: false
+})
+export class AppLandingComponent {
+  angularVersion = VERSION;
+
+  constructor() {}
+
+}
