+++ packages/angular/test/base/src/app/standalone/home-page/home-page.component.html
@@ -1,0 +1,161 @@
+<ion-header>
+  <ion-toolbar>
+    <ion-buttons slot="start">
+      <ion-back-button defaultHref="/"></ion-back-button>
+    </ion-buttons>
+    <ion-title>
+      Test App - Angular {{ angularVersion.major }} - Standalone
+    </ion-title>
+  </ion-toolbar>
+</ion-header>
+<ion-content>
+  <ion-list>
+    <ion-list-header>
+      <ion-label>Components</ion-label>
+    </ion-list-header>
+    <ion-item routerLink="/standalone/back-button">
+      <ion-label>
+        Back Button Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/button">
+      <ion-label>
+        Button Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/icon">
+      <ion-label>
+        Icon Test
+      </ion-label>
+    </ion-item>
+  </ion-list>
+
+  <ion-list>
+    <ion-list-header>
+      <ion-label>Navigation</ion-label>
+    </ion-list-header>
+    <ion-item routerLink="/standalone/nav">
+      <ion-label>
+        Nav Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/router-link">
+      <ion-label>
+        Router Link Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/router-outlet">
+      <ion-label>
+        Router Outlet Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/tabs">
+      <ion-label>
+        Tabs Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/tabs-basic">
+      <ion-label>
+        Tabs Basic Test
+      </ion-label>
+    </ion-item>
+  </ion-list>
+
+  <ion-list>
+    <ion-list-header>
+      <ion-label>Overlays</ion-label>
+    </ion-list-header>
+    <ion-item routerLink="/standalone/action-sheet-controller">
+      <ion-label>
+        Action Sheet Controller Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/menu-controller">
+      <ion-label>
+        Menu Controller Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/modal">
+      <ion-label>
+        Modal Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/overlay-controllers">
+      <ion-label>
+        Overlay Controllers Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/popover">
+      <ion-label>
+        Popover Test
+      </ion-label>
+    </ion-item>
+  </ion-list>
+
+  <ion-list>
+    <ion-list-header>
+      <ion-label>Value Accessors</ion-label>
+    </ion-list-header>
+    <ion-item routerLink="/standalone/value-accessors/checkbox">
+      <ion-label>
+        Checkbox Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/datetime">
+      <ion-label>
+        Datetime Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/input">
+      <ion-label>
+        Input Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/radio-group">
+      <ion-label>
+        Radio Group Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/range">
+      <ion-label>
+        Range Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/searchbar">
+      <ion-label>
+        Searchbar Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/segment">
+      <ion-label>
+        Segment Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/select">
+      <ion-label>
+        Select Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/textarea">
+      <ion-label>
+        Textarea Test
+      </ion-label>
+    </ion-item>
+    <ion-item routerLink="/standalone/value-accessors/toggle">
+      <ion-label>
+        Toggle Test
+      </ion-label>
+    </ion-item>
+  </ion-list>
+
+  <ion-list>
+    <ion-list-header>
+      <ion-label>Miscellaneous</ion-label>
+    </ion-list-header>
+    <ion-item routerLink="/standalone/providers">
+      <ion-label>
+        Providers Test
+      </ion-label>
+    </ion-item>
+  </ion-list>
+</ion-content>
